# docs/swagger-doc.yaml
openapi: 3.0.1
info:
  title: Tunemyles API
  description: >-
    This is a sample server Petstore server. You can find out more about Swagger
    at [http://swagger.io](http://swagger.io) or on
    [irc.freenode.net](irc.freenode.net)
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'http://localhost:5000/api/v2'
    description: Dev Server API URL
...
paths:
...
  /wallets/withdraw:
    post:
      tags:
        - wallets
      summary: Withdraw (Route All Roles)
      description: Withdraw (Route All Roles)
      operationId: withdraw
      requestBody:
        description: Request Data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
      x-codegen-request-body-name: body
...
  '/wallets/topup/{id}':
    get:
      tags:
        - wallets
      summary: Get Top Up Detail (Route All Roles)
      description: Get Top Up Detail (Route All Roles)
      operationId: topUpDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTopUpDetailResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
  '/wallets/withdraw/{id}':
    get:
      tags:
        - wallets
      summary: Get Withdraw Detail (Route All Roles)
      description: Get Withdraw Detail (Route All Roles)
      operationId: withdrawDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetWithdrawDetailResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
  /wallets/topup/users/all:
    get:
      tags:
        - wallets
      summary: Get All User Top Up (Admin Route)
      description: Get All User Top Up (Admin Route)
      operationId: allUserTopUp
      parameters:
        - name: orderDirection
          in: query
          schema:
            $ref: '#/components/schemas/OrderDirectionQuery'
        - name: orderBy
          in: query
          schema:
            allOf:
              - $ref: '#/components/schemas/OrderByQuery'
              - enum:
                  - created_at
                  - balance_request
                  - balance_transfer
        - name: search
          in: query
          description: 'Available search field balance_request, balance_transfer'
          schema:
            $ref: '#/components/schemas/SearchQuery'
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/PageQuery'
        - name: limit
          in: query
          schema:
            $ref: '#/components/schemas/LimitQuery'
        - name: status
          in: query
          schema:
            allOf:
              - $ref: '#/components/schemas/StatusQuery'
              - enum:
                  - '0'
                  - '1'
                  - '2'
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllUserTopUpResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
  /wallets/withdraw/users/all:
    get:
      tags:
        - wallets
      summary: Get All User Withdraw (Admin Route)
      description: Get All User Withdraw (Admin Route)
      operationId: allUserWithdraw
      parameters:
        - name: orderDirection
          in: query
          schema:
            $ref: '#/components/schemas/OrderDirectionQuery'
        - name: orderBy
          in: query
          schema:
            allOf:
              - $ref: '#/components/schemas/OrderByQuery'
              - enum:
                  - created_at
                  - balance_request
        - name: search
          in: query
          description: Available search field balance_request
          schema:
            $ref: '#/components/schemas/SearchQuery'
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/PageQuery'
        - name: limit
          in: query
          schema:
            $ref: '#/components/schemas/LimitQuery'
        - name: status
          in: query
          schema:
            allOf:
              - $ref: '#/components/schemas/StatusQuery'
              - enum:
                  - '0'
                  - '1'
                  - '2'
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllUserWithdrawResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
...
  '/wallets/topup/{id}/status':
    patch:
      tags:
        - wallets
      summary: Update Top Up Status (Admin Route)
      description: Update Top Up Status (Admin Route)
      operationId: updateTopUpStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Request Data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTopUpStatusBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
      x-codegen-request-body-name: body
  '/wallets/withdraw/{id}/status':
    patch:
      tags:
        - wallets
      summary: Update Withdraw Status (Admin Route)
      description: Update Withdraw Status (Admin Route)
      operationId: updateWithdrawStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Request Data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWithdrawStatusBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
      x-codegen-request-body-name: body
...
components:
  schemas:
    PageQuery:
      type: string
      default: '1'
    SearchQuery:
      type: string
      default: ''
    OrderByQuery:
      type: string
      default: created_at
    OrderDirectionQuery:
      type: string
      default: DESC
      enum:
        - ASC
        - DESC
    LimitQuery:
      type: string
      default: '10'
...
    StatusQuery:
      type: string
      default: '0'
    ApiResponse:
      type: object
      properties:
        code:
          type: number
        success:
          type: boolean
        messages:
          type: string
...
    WithdrawBody:
      required:
        - user_bank_id
        - balance_request
      type: object
      properties:
        user_bank_id:
          type: integer
        balance_request:
          type: integer
...
    GetTopUpDetailResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                id:
                  type: integer
                balance_request:
                  type: integer
                balance_transfer:
                  type: integer
                status:
                  type: integer
                user_name:
                  type: string
                created_at:
                  type: string
                  format: date-time
                updated_at:
                  type: string
                  format: date-time
    GetWithdrawDetailResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                id:
                  type: integer
                balance_request:
                  type: integer
                status:
                  type: integer
                bank_name:
                  type: string
                created_at:
                  type: string
                  format: date-time
                updated_at:
                  type: string
                  format: date-time
    GetAllUserTopUpResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  balance_request:
                    type: integer
                  balance_transfer:
                    type: integer
                  status:
                    type: integer
                  bank_name:
                    type: string
                  proof_label:
                    type: string
                    nullable: true
                  proof_url:
                    type: string
                    nullable: true
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
    GetAllUserWithdrawResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  balance_request:
                    type: integer
                  status:
                    type: integer
                  bank_name:
                    type: string
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
...
    UpdateTopUpStatusBody:
      type: object
      properties:
        status:
          type: integer
          minimum: 0
          maximum: 1
    UpdateWithdrawStatusBody:
      type: object
      properties:
        status:
          type: integer
          minimum: 0
          maximum: 1
...
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      name: authorization
      in: header